Coroutine Guide - Cancellation and Timeouts
> Cancelling coroutine execution
