--- src/main/kotlin/toc2/cancellation_and_timeouts.kt
@@ -1,17 +1,0 @@
-package toc2
-
-import kotlinx.coroutines.*
-
-fun main() = runBlocking {
-    val job = launch {
-        repeat(1000) { i ->
-            println("job: I'm sleeping $i ...")
-            delay(500L)
-        }
-    }
-    delay(1300L) // delay a bit
-    println("main: I'm tired of waiting!")
-    job.cancel() // cancels the job
-    job.join() // waits for job's completion
-    println("main: Now I can quit.")
-}
